Dish service and repository final

+++ src/main/java/ru/agorbunov/restaurant/service/DishService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import ru.agorbunov.restaurant.model.Dish;
 import ru.agorbunov.restaurant.repository.DishRepository;
+import ru.agorbunov.restaurant.repository.UserRepository;
 
 import java.util.List;
 
@@ -15,17 +16,16 @@ public class DishService {
 
     private final DishRepository repository;
 
+    private final UserRepository userRepository;
+
     @Autowired
-    public DishService(DishRepository repository) {
+    public DishService(DishRepository repository, UserRepository userRepository) {
         this.repository = repository;
+        this.userRepository = userRepository;
     }
 
-    public Dish create(Dish dish) {
-        Assert.notNull(dish, "dish must not be null");
-        return repository.save(dish);
-    }
-
-    public void delete(int id) {
+    public void delete(int id, int userId) {
+        userRepository.checkIsAdmin(userId);
         checkNotFoundWithId(repository.delete(id), id);
     }
 
@@ -37,7 +37,9 @@ public class DishService {
         return repository.getAll();
     }
 
-    public void update(Dish dish) {
+
+    public void update(Dish dish, int userId) {
+        userRepository.checkIsAdmin(userId);
         Assert.notNull(dish, "dish must not be null");
         repository.save(dish);
     }
